reshaping the code
